Split legacy code and new code into partial classes and partial resources
