+++ ScanbotBarcodeSDKExample/BarcodeSDK.NET.Droid.Example/BarcodeSDK.NET.Droid.Example.csproj
@@ -13,6 +13,8 @@
         <Platforms>x64;ARM64</Platforms>
         <CreatePackage>false</CreatePackage>
         <WarningsAsErrors>NU1603</WarningsAsErrors>
+        <!-- Comment this out to see RTU v1 in the example -->
+        <!-- <DefineConstants>$(DefineConstants);LEGACY_EXAMPLES</DefineConstants> -->
     </PropertyGroup>
     <ItemGroup>
          <PackageReference Include="Scanbot.Xamarin.ImagePicker" Version="0.0.3-beta.1" />
